fix #106 backend crash after contact creation
On initial project creation, there was no empty contact array created and added to the database document for the project. Then when trying to add a new contact to the non existent array, the crash occured.

This commit fixes this issue and also fixes the same issue for excavations and sections.